feat: Add fetchData function for fetching page data from Sanity
This commit adds a new function called `fetchData` to the `client.ts` file in the `sanity/lib` directory. The `fetchData` function retrieves page data from Sanity by executing a query to fetch pages with their IDs, names, and parent IDs. This function will be used in the `Sidebar` component to fetch folders data. This change enhances the functionality of the application by providing a centralized method for fetching page data from Sanity.